% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/common_tools.R
\name{add_notches}
\alias{add_notches}
\title{Add notch-like gaussian snippets to an existing signal x}
\usage{
add_notches(x, pos, sd = 0.01 * length(x), amplitude = 1)
}
\arguments{
\item{x}{[1,N] numeric, Original data}

\item{pos}{[1,m] integer, Positions to add notches to}

\item{sd}{[1,1] numeric, (optional) Desired width of the Gaussian notch}

\item{amplitude}{[1,1] numeric, (optional) Desired amplitude for the notches}
}
\value{
[1,N] numeric, Modified signal with notches
}
\description{
Add notch-like gaussian snippets to an existing signal x
}

