% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/cocoreg.R
\name{apply_dc_meta}
\alias{apply_dc_meta}
\title{Apply extracted properties of a data collection to a data collection (restore)}
\usage{
apply_dc_meta(df_list, meta)
}
\arguments{
\item{df_list}{list of data.frames, The data collection to process}

\item{meta,}{list, Output of get_dc_meta()}
}
\value{
A list of data.frames, the data collection with updated metadata
}
\description{
Apply extracted properties of a data collection to a data collection (restore)
}

