% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/cocoreg.R
\name{average_R2_dflst}
\alias{average_R2_dflst}
\title{Computes the R^2 (variance explained) between two lists of data.frames}
\usage{
average_R2_dflst(df_orig_lst, df_est_lst)
}
\arguments{
\item{df_orig_lst}{List of original data.frames}

\item{df_est_lst}{List of estimated data.frames}
}
\value{
Returns a data.frame with R2 values, one value for each data set and variable. Molten/long format.
}
\description{
Computes the R^2 (variance explained) between two lists of data.frames
}

