% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/common_tools.R
\name{BGFA_joint_info}
\alias{BGFA_joint_info}
\title{Project BGFA components common to all datasets back to the original space}
\usage{
BGFA_joint_info(model, threshold = 0.001)
}
\arguments{
\item{model}{Output of CCAGFA::GFA()}

\item{threshold}{[1,1] double, GFA model trimming threshold}
}
\value{
A list of data.frames, Original data reconstructed using only
those latent components that are active in all datasets
}
\description{
Project BGFA components common to all datasets back to the original space
}

