% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/cocoreg.R
\name{cocoreg_by_path}
\alias{cocoreg_by_path}
\title{Compute D_joint for dataset i separately for all paths
Can be used to study path variability}
\usage{
cocoreg_by_path(data_orig, ccr, ds_ind)
}
\arguments{
\item{data_orig}{list of data.frames, Original data collection}

\item{ccr}{list, output of cocoreg(data_orig)}

\item{ds_ind}{integer, index of the dataset to process}
}
\value{
A list of data.frames, D_joint corresponding to each path that ends
        at 'ds_ind' i.e. paths defined by ccr$paths[,ds_ind].
        Dimensions of the matrices are the same as for the data.frames
        in data_orig.
}
\description{
Compute D_joint for dataset i separately for all paths
Can be used to study path variability
}
\examples{
\dontrun{
ccr <- cocoreg(data_list)
jibp <- cocoreg_by_path(ccr, 1)
}
}

