% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/cocoreg.R
\name{compose}
\alias{compose}
\title{Calculate the composition formed by applying all functions
in the given path to a dataset.}
\usage{
compose(x, path, mappings)
}
\arguments{
\item{x}{A data frame or vector}

\item{path}{A list describing the path.}

\item{mappings}{A list containing the mapping functions described in the path.
                 Usually a list of all M*M-M available mappings between the M
                 data sets.}
}
\value{
A vector containing the result of the composition.
}
\description{
Calculate the composition formed by applying all functions
in the given path to a dataset.
}

