% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/cocoreg.R
\name{compose_all}
\alias{compose_all}
\title{Calculate the average of the composition formed by applying all functions
in all possible paths to a dataset.}
\usage{
compose_all(x, paths, mappings)
}
\arguments{
\item{x}{A list of data frames.}

\item{paths}{A list of list with paths.}

\item{mappings}{A list containing the mapping functions described in the path.}
}
\value{
A vector containing the average composition from all the paths.
}
\description{
Calculate the average of the composition formed by applying all functions
in all possible paths to a dataset.
}

