% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/cocoreg.R
\name{create_mappings}
\alias{create_mappings}
\title{Generate all possible pairwise mappings between the given multivariate
datasets.}
\usage{
create_mappings(data, mapping_function = mapping_lm)
}
\arguments{
\item{data}{A list of data.frames (the datasets)}

\item{mapping_function}{(optional) Default is \code{mapping_lm}.}
}
\value{
A named list containing the pairwise mapping functions.
}
\description{
The following naming convention is used in the output:
'1-2' means '1' mapped to '2', i.e.,, '2' explained by '1'.
}

