% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/synthetic_datasets.R
\name{create_syn_data_uvar}
\alias{create_syn_data_uvar}
\title{A dollection with unrelated variables}
\usage{
create_syn_data_uvar()
}
\value{
A list with elements
\item{data}{Data collection as a list of data.frames}
\item{Z_all}{Signals shared by all datasets in the collection}
\item{Z_sub}{Signals not shared by all datasets}
\item{W_all}{Mixing weights for Z_all}
\item{W_sub}{Mixing weights for Z_sub}
\item{E}{Noise}
\item{var.coef}{Noise multiplication factor used}
}
\description{
A dollection with unrelated variables
}

