% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/synthetic_datasets.R
\name{create_syndata_pwl}
\alias{create_syndata_pwl}
\title{A non-linear data collection using piecewise linearity}
\usage{
create_syndata_pwl()
}
\value{
A list with elements
\item{data}{Data collection as a list of data.frames}
\item{Z}{Signals used}
\item{W}{Mixing matrix used}
\item{Z_all}{Signals shared by all datasets in the collection}
\item{Z_sub}{Signals not shared by all datasets}
\item{W_all}{Mixing weights for Z_all}
\item{W_sub}{Mixing weights for Z_sub as a list of matrices, one matrix per dataset}
\item{E}{Noise}
\item{var.coef}{Noise multiplication factor used}
}
\description{
A non-linear data collection using piecewise linearity
}

