% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/synthetic_datasets.R
\name{create_Z_linear}
\alias{create_Z_linear}
\title{Contains functions to create synthetic datasets with different properties.
The create_syn_data_*() functions follow the scheme:
"total variation = shared_by_all + shared_by_subset + noise"
Create signals}
\usage{
create_Z_linear(N, decorrelate = T)
}
\arguments{
\item{N}{Number of observations in data as integer}

\item{decorrelate}{(optional) Should the variables be decorrelated?}
}
\value{
A [N,3] matrix of signals
}
\description{
Contains functions to create synthetic datasets with different properties.
The create_syn_data_*() functions follow the scheme:
"total variation = shared_by_all + shared_by_subset + noise"
Create signals
}

