% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/common_tools.R
\name{cshift}
\alias{cshift}
\title{Circularly shift vector elements}
\usage{
cshift(x, by)
}
\arguments{
\item{x}{[1,N] numeric, A vector}

\item{by}{[1,1] integer, How many positions to shift.
       by > 0 -> shift to right
       by = 0 -> no shift
       by < 0 -> shift to left}
}
\value{
[1,N] numeric, Circularly shifted signal
}
\description{
Circularly shift vector elements
}

