% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/cocoreg.R
\name{dc_variability}
\alias{dc_variability}
\title{Compute ds_variability for all datasets in a data collection}
\usage{
dc_variability(data, ccr)
}
\arguments{
\item{data}{Unprocessed dataset as a list of data.frames}

\item{ccr}{Processed dataset as a list of data.frames, output of cocoreg()$data}
}
\value{
Path variability as a data.frame
}
\description{
Compute ds_variability for all datasets in a data collection
}

