% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/common_tools.R
\name{df_scale}
\alias{df_scale}
\title{Apply scale on a numeric data.frame}
\usage{
df_scale(df, ...)
}
\arguments{
\item{df}{data.frame, A numeric data.frame to process}

\item{...}{arguments to scale()}
}
\value{
data.frame, A scaled data.frame with attributes preserved
}
\description{
Apply scale on a numeric data.frame
}

