% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/common_tools.R
\name{dflst2array}
\alias{dflst2array}
\title{Catenate a list of data.frames to a matrix along dim}
\usage{
dflst2array(df_lst, dim = 2)
}
\arguments{
\item{df_lst}{[1,m] list of data.frames, A list of data.frames to process}

\item{dim}{[1,1] int, Dimension to apply over}
}
\value{
A matrix with elements of df_lst converted to matrix and catenated along dim.
}
\description{
Catenate a list of data.frames to a matrix along dim
}

