% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/common_tools.R
\name{dflst2df}
\alias{dflst2df}
\title{Catenate a list of data.frames vertically to a single data.frame}
\usage{
dflst2df(df_lst, id_var_name = "dataset")
}
\arguments{
\item{df_lst}{[1,m] list of data.frames, A list of data.frames to process}

\item{id_var_name}{string, Column name for the dataset id variable}
}
\value{
A data.frame with elements of df_lst catenated vertically, An extra column with dataset id
is added.
}
\description{
Assumes equal variables for all datasets!
Ouput has columns: <variables>, "dataset"
Preserves list element names in column "dataset".
For a more generic approach see dflst2dfmelt (uses reshape::melt(df_lst))
}

