% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/common_tools.R
\name{dflst2dfmelt}
\alias{dflst2dfmelt}
\title{Combine a list of data.frames to a single molten data.frame}
\usage{
dflst2dfmelt(df_lst)
}
\arguments{
\item{df_lst}{[1,m] list of data.frames, A list of data.frames to process}
}
\value{
A data.frame with elements of df_lst combined using reshape::melt().

Extra columns:

\describe{
\item{dataset}{dataset name}
\item{obs}{running observation index (time)}
}
}
\description{
Output maximally "molten" with columns "dataset", "obs", "variable", "value"
Preserves list element names in column "dataset".
}

