% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/common_tools.R
\name{dflst_add_ds}
\alias{dflst_add_ds}
\title{Add a data.frame (dataset) to a list of data.frames (datasets)}
\usage{
dflst_add_ds(dflst, df, dsname)
}
\arguments{
\item{dflst}{[1,m] list of data.frames, A list of data.frames}

\item{df}{data.frame, Data frame to add}

\item{dsname}{string, Dataset name for the data.frame to add}
}
\value{
A list of data.frames, A new list of data.frames with one new dataset in the end
}
\description{
Add a data.frame (dataset) to a list of data.frames (datasets)
}

