% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/common_tools.R
\name{dflst_dsnames2varnames}
\alias{dflst_dsnames2varnames}
\title{Append dataset names to variable names of the respective dataset}
\usage{
dflst_dsnames2varnames(dflst, sep = "_")
}
\arguments{
\item{dflst}{[1,m] list of data.frames, A list of data.frames to process}

\item{sep}{string, Separator to use}
}
\value{
A [1,m] list of data.frames with modified variable names.
}
\description{
Append dataset names to variable names of the respective dataset
}

