% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/common_tools.R
\name{dflst_pca}
\alias{dflst_pca}
\title{Apply PCA to the data after catenating data.frames horizontally}
\usage{
dflst_pca(df_lst, center = F, scale = F)
}
\arguments{
\item{df_lst}{[1,m] list of data.frames, A list of data.frames to process}

\item{center}{boolean, TRUE -> center data, FALSE -> do nothing}

\item{scale}{boolean, TRUE -> scale data, FALSE -> do nothing}
}
\value{
A list with elements:
\item{pcdf:}{data.frame, PCA components (prcomp()$x)}
\item{model:}{list, Output of prcomp()}
}
\description{
Apply PCA to the data after catenating data.frames horizontally
}

