% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/common_tools.R
\name{dl_remove_NA}
\alias{dl_remove_NA}
\title{Remove rows with NA values from a list of data.frames}
\usage{
dl_remove_NA(df_lst)
}
\arguments{
\item{df_lst}{[1,m] list of data.frames, A list of data.frames to process}
}
\value{
[1,m] list of data.frames, Data without rows that contain NA,
same rows removed from all data.frames in the input list.
}
\description{
Same rows removed from all data frames in the list.
}

