% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/synthetic_datasets.R
\name{dl_scale}
\alias{dl_scale}
\title{Run scale() on a list of data.frames}
\usage{
dl_scale(dl, ...)
}
\arguments{
\item{dl}{A list of data.frames}

\item{...}{Additional arguments to be passed on to scale}
}
\value{
A list of data.frames that have been processed using scale() and
converted back to data.frame
}
\description{
Run scale() on a list of data.frames
}

