% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/cocoreg.R
\name{ds_variability}
\alias{ds_variability}
\title{Compute variability_components for a dataset}
\usage{
ds_variability(data, ccr, ds_ind)
}
\arguments{
\item{data}{Unprocessed dataset as a list of data.frames}

\item{ccr}{Processed dataset as a list of data.frames, output of cocoreg()$data}

\item{ds_ind}{Starting dataset of the set of paths to analyze as [1,1] integer}
}
\value{
Path variability as a data.frame
}
\description{
Note: might not work for cyclic ccr <TODO>
}

