% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/cocoreg.R
\name{generate_mapping_function}
\alias{generate_mapping_function}
\title{Generate a mapping function between two datasets}
\usage{
generate_mapping_function(method = lm)
}
\arguments{
\item{method}{A function to be used in the mapping. A function object.}
}
\value{
Returns a function object that does the mapping between two datasets i.e.
from dataset 1 to dataset 2.
}
\description{
Generate a mapping function between two datasets,
using some method, such as linear regression (lm),
or some classifier such as a random forest (randomForest).

Wraps the function as well as data into a single object.
}

