% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/cocoreg.R
\name{generate_paths_noncyclic}
\alias{generate_paths_noncyclic}
\title{Generate non-cyclic paths}
\usage{
generate_paths_noncyclic(ind, n, sample_paths = FALSE)
}
\arguments{
\item{ind}{The starting dataset}

\item{n}{The number of datasets.}

\item{sample_paths}{boolean, If FALSE, all possible paths are generated.
       If TRUE one path per ending point is selected.}
}
\value{
A list of lists containing the paths.
}
\description{
From a set of n numbers, generate all possible paths starting from
and ending on a given number.
}

