% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/cocoreg.R
\name{get_dc_meta}
\alias{get_dc_meta}
\title{Extract important properties of data collection}
\usage{
get_dc_meta(df_list, type = "current")
}
\arguments{
\item{df_list}{list of data.frames, The data collection to process}

\item{type}{string, If 'current' then data collection metadata is collected from the
       data collection itself. If 'original' then metadata is collected from special attributes.}
}
\value{
A list with elements:
\item{$dcnames:}{Dataset names}
\item{$dcdimnames:}{A list of dataset dimension names for each dataset}
}
\description{
Extract important properties of data collection
}

