% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/common_tools.R
\name{get_range_datalist}
\alias{get_range_datalist}
\title{Get [min, max] of a list of numeric objects}
\usage{
get_range_datalist(dataList)
}
\arguments{
\item{dataList}{[1,m] list of numeric objects}
}
\value{
[1,2] double, [min, max] of the input
}
\description{
Get [min, max] of a list of numeric objects
}

