% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/cocoreg.R
\name{get_starting_dataset}
\alias{get_starting_dataset}
\title{Helper function to get the starting dataset based on
a path.}
\usage{
get_starting_dataset(p)
}
\arguments{
\item{p}{A path.}
}
\value{
A number indicating the starting dataset
}
\description{
Helper function to get the starting dataset based on
a path.
}

