% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/common_plots.R
\name{ggcompare_dclst}
\alias{ggcompare_dclst}
\title{Compare data collections variable by variable}
\usage{
ggcompare_dclst(dclst)
}
\arguments{
\item{dclst}{A (named) list of data collections i.e. a list of lists of data.frames (see examples)}
}
\value{
Returns a ggplot object (which is by default printed if not assigned to variable)
}
\description{
Compare data collections variable by variable
}

