% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/common_plots.R
\name{ggplot_df}
\alias{ggplot_df}
\title{Plotting data.frame using ggplot}
\usage{
ggplot_df(df, titlestr = NULL, ylabstr = NULL, ylim = NULL,
  color = NULL, linetype = NULL, logy = F)
}
\arguments{
\item{df}{A data.frame to plot}

\item{titlestr}{(optional) Title of plot as string}

\item{ylabstr}{(optional) Y-axis label as string}

\item{ylim}{(optional) y-axis limits as [1,2] numeric, passed on to dfplot() as 'ylim'}

\item{color}{(optional) Input for manual color scale}

\item{linetype}{(optional) Input for manual linetype scale}

\item{logy}{(optional) Should y-axis be logarithmic? A boolean value.}
}
\value{
Returns a ggplot2 object
}
\description{
Plotting data.frame using ggplot
}
\examples{
\dontrun{
dc <- create_syn_data_toy()
ggplot_df(dc$data[[1]])
}
}

