% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/common_plots.R
\name{ggplot_dflst}
\alias{ggplot_dflst}
\title{Plot a list of data.frames using ggplot2}
\usage{
ggplot_dflst(dflst, ncol = 1, plot = T, plotfun = ggplot_df, ...)
}
\arguments{
\item{dflst}{A list of datasets as a list of data.frames}

\item{ncol}{(optional) Number of columns in final plot}

\item{plot}{(optional) Plot or not: if TRUE produces a plot else returns a list of ggplot objects}

\item{plotfun}{(optional) Function used to plot a data.frame (one panel in final plot)}

\item{...}{(optional) Additional parameters passed on to plotfun}
}
\value{
Produces a plot to the active graphics device or returns a list of ggplot objects
}
\description{
Plot a list of data.frames using ggplot2
}
\examples{
\dontrun{
dc <- create_syn_data_toy()
ggplot_dflst(dc$data)
}
}

