% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/common_tools.R
\name{make_data_gauss_2d}
\alias{make_data_gauss_2d}
\title{Make 2D gauss data (maybe obsolete)}
\usage{
make_data_gauss_2d(n, var, angle_deg, scale = T, seed = 42)
}
\arguments{
\item{n}{[1,1] int, Number of observations}

\item{var}{[1,2] numeric, Variances}

\item{angle_deg}{[1,1] numeric, Rotation angle}

\item{scale}{boolean, Scale data? T -> scale, F -> do not scale}

\item{seed}{[1,1] int, Random seed}
}
\value{
Matrix of 2D gaussian data
}
\description{
Make 2D gauss data (maybe obsolete)
}

