% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/cocoreg.R
\name{mapping_pcr}
\alias{mapping_pcr}
\title{Define a mapping function using pls::pcr}
\usage{
mapping_pcr(data1, data2)
}
\arguments{
\item{data1}{data.frame, Dataset 1, the independent variables}

\item{data2}{data.frame, Dataset 2, the dependent variables}
}
\value{
Returns a function object that does the mapping between two datasets.
}
\description{
Define a mapping function using pls::pcr
}

