% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/cocoreg.R
\name{mapping_rf}
\alias{mapping_rf}
\title{Mapping randomForest}
\usage{
mapping_rf(data1, data2)
}
\arguments{
\item{data1}{data.frame, Dataset 1, the independent variables}

\item{data2}{data.frame, Dataset 2, the dependent variables}
}
\description{
Mapping randomForest
}

