% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/common_plots.R
\name{nplst_reorder_grid}
\alias{nplst_reorder_grid}
\title{Reorders a nested list of ggplots}
\usage{
nplst_reorder_grid(p_list, ncol)
}
\arguments{
\item{p_list}{A list of lists of ggplots}

\item{ncol}{Target number of columns, integer value}
}
\value{
A reordered and flattened version of input list as a list of ggplot2 objects
}
\description{
Reorders a nested list of ggplots to ncol columns prior to calling grid.arrange()
Note: p_list is a list of lists of ggplots.
p_list = list(p_list1, p_list2,...)
}

