% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/cocoreg.R
\name{pathify}
\alias{pathify}
\title{Create a path}
\usage{
pathify(ind, p, cyclic = FALSE)
}
\arguments{
\item{ind}{[1,1] int, The starting point.}

\item{p}{[1,m] int, A set of numbers through which the path must go,
excluding the starting point.}

\item{cyclic}{boolean, If TRUE, the path leads back to the starting point.
       If FALSE, the path stops at last element of p.}
}
\value{
A list containing the path definition.
}
\description{
Given a set of numbers and a starting point, create a path from
the starting point through each point. If cyclic = FALSE, the path
stops at p[length(p)] and if cyclic = TRUE the patch goes back to ind.
}
\keyword{internal}

