% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/common_tools.R
\name{PCA_cocoreg_interface}
\alias{PCA_cocoreg_interface}
\title{PCA projection using cocoreg interface}
\usage{
PCA_cocoreg_interface(df_list, prc_th = 0.9)
}
\arguments{
\item{df_list}{[1,m] list of data.frames, Input data to GFA in COCOREG format}

\item{prc_th}{[1,1] double, Threshold in precentage of cumulative variance explained
       PCA components are included until cumulative explained variance reaches prc_th.}
}
\value{
A list with elements:
\item{$data}{[1,m] list of data.frames, Original data reconstructed using only
        those latent components that are active in all datasets}
\item{$dataid}{string, Dataset identifier string}
\item{$method}{string, Analysis method identifier string}
\item{$wall_time_taken}{[1,1] double, Time taken to run the analysis in seconds}
}
\description{
PCA projection using cocoreg interface
}

