% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/common_tools.R
\name{repmat}
\alias{repmat}
\title{Replicate matrix to create a larger one}
\usage{
repmat(X, m, n)
}
\arguments{
\item{X}{A [I,J] matrix or J element vector, Matrix used as such, vector coerced to a row
matrix with dim(X)=[1,J].}

\item{m}{[1,1] integer, Replication count vertically}

\item{n}{[1,1] integer, Replication count horizontally}
}
\value{
[m*I,n*J] matrix, Replicated data
}
\description{
From:
http://haky-functions.blogspot.fi/2006/11/repmat-function-matlab.html (accessed 27.3.2015)
}

