% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/common_tools.R
\name{rotation_matrix}
\alias{rotation_matrix}
\title{A rotation matrix}
\usage{
rotation_matrix(angle_deg)
}
\arguments{
\item{angle_deg}{[1,1] numeric, Angle in degrees}
}
\value{
[2,2] matrix, Rotation matrix for making angle_deg 2D rotation
}
\description{
A rotation matrix
}

