% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/common_tools.R
\name{se}
\alias{se}
\title{Standard error of mean}
\usage{
se(x, na.rm = T)
}
\arguments{
\item{x}{[1,M] numeric, data vector}

\item{na.rm}{procedure for NA's, passed on to sd(), default: na.rm = T}
}
\value{
[1,1] numeric, standard error of mean
}
\description{
Standard error of mean
}

