% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/cocoreg.R
\name{svm_sigmoid}
\alias{svm_sigmoid}
\title{SVM using sigmoid kernel}
\usage{
svm_sigmoid(...)
}
\arguments{
\item{...}{Further arguments passed on to e1071::svm()}
}
\description{
SVM using sigmoid kernel
}

