% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/common_tools.R
\name{to_unit_vec}
\alias{to_unit_vec}
\title{Make vector of unit norm}
\usage{
to_unit_vec(x)
}
\arguments{
\item{x}{[1,m] double, A vector of data}
}
\value{
[1,m] double, Same vector normalized to unit Euclidean norm
}
\description{
Make vector of unit norm
}

