% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/cocoreg.R
\name{validate_data}
\alias{validate_data}
\title{Validate a data collection for use with cocoreg}
\usage{
validate_data(df_list)
}
\arguments{
\item{df_list}{list of data.frames, The data collection to validate}
}
\value{
A list of data.frames that conform to the requirements
}
\description{
Check that the data collection has all the required properties.
}

