% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/cocoreg.R
\name{vector_variability}
\alias{vector_variability}
\title{Compute "variance" of the vectors var()}
\usage{
vector_variability(vec_lst, mean_vec = apply(abind::abind(vec_lst, along = 2),
  1, mean))
}
\arguments{
\item{vec_lst}{Data to process as a list of numeric vectors}

\item{mean_vec}{(optional) Desired mean vector as a numeric vector}
}
\value{
Variance of data values around mean as a numeric matrix
}
\description{
Compute "variance" of the vectors var()
}

