% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/common_tools.R
\name{wrapper_BGFA}
\alias{wrapper_BGFA}
\title{Run BGFA by Klami et al using data format conventions of this repo}
\usage{
wrapper_BGFA(df_list, K = 8, Nrep = 2)
}
\arguments{
\item{df_list}{[1,m] list of data.frames, Input data to GFA in COCOREG format}

\item{K}{[1,1] int, The (maximum) number of components; should be high enough to capture all of the components. This can be recognized by at least a few of the components being shut down}

\item{Nrep}{[1,1] int, Number of random initializatin used for learning the model}
}
\value{
A list, The output of CCAGFA::GFA()
}
\description{
Run BGFA by Klami et al using data format conventions of this repo
}

